// garageuberweb/src/app/models/register_garage.ts
export interface RegisterGarage{
  id: number;
  garageName: string;
  garageLocation: string;
  garagePhone: string;
  garageEmail: string;
  garageOwnerName: string;
  garageLogoImageURL: string;
  service1: string;
  service1Charges: number;
  service2: string;
  service2Charges: number;
  service3: string;
  service3Charges: number;
  service4: string;
  service4Charges: number;
  service5: string;
  service5Charges: number;
  garageCode: string;
}